Sublime Text 3: modify code for retrieving current selections
The code which sets current selections in ParinferApplyCommand fails
in Sublime Text 3, due to use of a new API in Sublime Text 4 that is
not backward compatible.

This change maintains support for Sublime Text 3.

It is intended for personal use. The official version is designed and
maintained for Sublime Text 4.