Sublime Text 3: view clones (and buffers) are not available
Though the official documentation shows 'clones()' as available in ST3,
it is not. Buffers are clearly marked as not available. This means that
there is no way to be sure whether a buffer has multiple views, and the
on_close code can never do anything, so it has been removed.

The error is only visible with the ST console open on closing a view.