Support enabling Parinfer when saving a new file
Previously new files would not be enabled due to being unable to
determine whether they were a supported syntax. After save, the user
would need to manually enable Parinfer.

This change checks the type post-save so that on supported types
Parinfer is now automatically enabled.

Implementation is primitive: the on_post_save method calls the on_load
method so that the file acts the same as if closed and reopened.
Additionally, the check will run on saving any file, not only new files.
Ideally this would only happen on saving a new file.