Support enabling Parinfer when saving a new file
Previously new files would not be enabled due to being unable to
determine whether they were a supported syntax. After save,
the user would need to manually enable Parinfer.

This change checks whether to enable post-save so that on supported
types Parinfer is now automatically enabled.

Introduces a new status and renames the primary status collection.
This is so that on save we can check
- is Parinfer already enabled? (nothing to do)
- has the user explicitly disabled Parinfer
  - do not override their intent

This is done because I could not find a way to determine whether
the file was a new file or one previously loaded from storage.
I tried on_pre_save but it did not offer anything useful.
So for existing files we want to obey the user's wishes.

Note: if the user has never previously enabled Parinfer, a save of
      any file will most probably enable it. This may need to be
      addressed.

Renames should_start to is_enabled_for_filetype to better express
intent and allow creation of a new should_start method which checks
current Parinfer status to see whether it should be automatically
enabled.

This is used in on_load and on_post_save.