Enable on first opening Sublime Text
Due to how Sublime Text loads, plugins are not necessarily enabled when
a file which was open on last quit is reopened on opening Sublime Text.

This leverages the on_init method which runs when a plugin has been
loaded. It calls on_load so that behaviour is consistent with manually
opening a file.